yoctogram-app-main/pyproject.toml
[tool.poetry]
name = "yoctogram"
version = "0.1.0"
description = "A barebones image-sharing web application to be used as an example for cloud deployments"
authors = ["Aditya Saligrama <aditya@saligrama.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.21"
alembic = "^1.12.1"
tenacity = "^8.2.3"
pydantic-settings = "^2.0.3"
sqlalchemy-utils = "^0.41.1"
pydantic = {extras = ["email"], version = "^2.4.2"}
boto3 = "^1.29.0"
boto3-stubs-lite = {extras = ["essential"], version = "^1.29.3"}
mypy-boto3-secretsmanager = "^1.29.0"
freezegun = "^1.2.2"
pydantic-core = "^2.20.1"
botocore = "^1.34.150"
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"
python-ulid = "^2.7.0"
redis = "^5.0.8"
structlog = "^24.4.0"
asgi-correlation-id = "^4.3.2"
rich = "^13.7.1"
joserfc = "^1.0.0"

[tool.poetry.group.dev.dependencies]
deptry = "^0.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
